pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "wavyopen"
requires-python = ">=3.10"
version = "0.3.2"
description = "A package for processing/calibrating/validating wave measurements and wave model output."
license = "MIT"
license-files = ["LICENSE"]

authors = [
  { name = "Patrik Bohlinger", email = "patrikb@met.no" }]

dependencies = [
  "matplotlib>=3.5",
  "numpy>=1.20",
  "scipy>=1.9",
  "netCDF4>=1.6",
  "pyyaml>=6",
  "pyproj>=2.3",
  "cmocean>=3",
  "joblib>=1.4",
  "pandas>=2.2",
  "xarray>=2024",
  "pyresample>=1.30",
  "geojson>=2.5",
  "Cartopy>=0.20",
  "pyGAM>=0.9",
  "roaring-landmask>=0.7",
  "copernicusmarine>=2.0",
  "xdg",
  "tqdm>=4",
  "python-dotenv",
  "click",
  "scikit-learn>=1.6"
  ]

[project.scripts]
wavyDownload = "wavy.apps.wavyDownload:main"
wavyCFG = "wavy.apps.wavyCFG:main"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--benchmark-disable"
testpaths = [
  "tests",
]

markers = [
  "need_credentials:need credentials to run"
]

[tool.pylint.master]
extension-pkg-whitelist="netCDF4,numpy,scipy,roaring_landmask"

[tool.pylint.messages_control]
disable = "all"
enable = "F,unreachable,duplicate-key,unnecessary-semicolon,global-variable-not-assigned,unused-variable,unused-wildcard-import,binary-op-exception,bad-format-string,anomalous-backslash-in-string,bad-open-mode,E0001,E0011,E0012,E0100,E0101,E0102,E0103,E0104,E0105,E0107,E0108,E0110,E0111,E0112,E0113,E0114,E0115,E0116,E0117,E0118,E0202,E0203,E0211,E0213,E0236,E0237,E0238,E0239,E0240,E0241,E0301,E0302,E0303,E0401,E0402,E0601,E0602,E0603,E0604,E0611,E0632,E0633,E0701,E0702,E0703,E0704,E0710,E0711,E0712,E1003,E1101,E1102,E1111,E1120,E1121,E1123,E1124,E1125,E1126,E1127,E1128,E1129,E1130,E1131,E1132,E1133,E1134,E1135,E1137,E1138,E1139,E1200,E1201,E1205,E1206,E1300,E1301,E1302,E1303,E1304,E1305,E1306,E1310,E1700,E1701"

